feat(mobile): reels, watch, video detail and upload screens
Reels: TikTok-style vertical PageView (GET /reels), autoplay active
page only, like/comment/mute overlay, comments bottom sheet. Unlike
web's Reels page (which only fakes the like button with local state
and never calls the API), this wires the real like endpoint since
it's already built out for Watch.

Watch: continue-watching row + recommended/trending/following tabs
backed by WatchNotifier's eager four-list load.

Video detail: shared by both Watch and Reels (a reel id is a video
row), video_player playback with a play/pause overlay, full
reaction picker (reused from posts' ReactionButton widget) matching
web's video player, comment thread with edit/delete for own comments.

Upload: title/description + image_picker's video mode (no new
picking dependency needed), two-step upload-then-create, isReel flag
routes to Watch or Reels.

Wired into the app shell: Reels/Watch icons on the feed app bar,
GoRoutes for /reels, /watch, /videos/upload.

Tests: 17 new unit tests across ReelsNotifier, WatchNotifier, and
VideoDetailNotifier (mocktail), plus 4 for StoriesNotifier landed in
the previous commit -- 21 new tests total, all passing. flutter
analyze clean.

+++ mobile/lib/features/posts/presentation/screens/feed_screen.dart
@@ -149,7 +149,7 @@ class _FeedScreenState extends ConsumerState<FeedScreen> {
         padding: const EdgeInsets.all(12),
         // +1 for the leading StoriesBar header item, on top of the feed items
         // and the trailing loader.
-        itemCount: 1 + feed.items.length + (feed.hasMore ? 1 : 0),
+        itemCount: (1 + feed.items.length + (feed.hasMore ? 1 : 0)) as int,
         itemBuilder: (context, index) {
           if (index == 0) return const StoriesBar();
           final itemIndex = index - 1;
